changed model fields to PositiveSmallIntegerField

+++ kyc/models.py
@@ -7,8 +7,8 @@ class Story(models.Model):
     title = models.CharField(max_length=256, unique=True)
     # place = models.ForeignKey(Place)
 
-    year = models.IntegerField()
-    month = models.IntegerField(null=True)
-    day = models.IntegerField(null=True)
+    year = models.PositiveSmallIntegerField()
+    month = models.PositiveSmallIntegerField(null=True)
+    day = models.PositiveSmallIntegerField(null=True)
 
     summary = models.CharField()
